use poetry,  small fixes, examples, scaling
