+++ README.md
@@ -1,0 +1,23 @@
+# Generate patterns from repetition in lyrics
+
+Bohemian Rhapsody by Queen
+
+<img src="https://user-images.githubusercontent.com/24855949/186299829-386c3c53-3c7f-463a-9b3a-466d4ecdecf3.png" width="400" />
+
+Three Hits by Indigo Girls
+
+<img src="https://user-images.githubusercontent.com/24855949/186299139-115d18bd-200a-4d33-b30d-d707480bd740.png" width="400" />
+
+Around the World by Daft Punk (truncated)
+
+<img src="https://user-images.githubusercontent.com/24855949/186299187-359266ea-3a9f-4b95-89a6-f607e8533e27.png" width="400" />
+
+# Usage
+
+Ensure you have [Poetry](https://python-poetry.org/) installed
+
+```
+poetry run generate ./lyricstext.txt
+```
+
+This saves a file `lyricstext.txt.png` next to the input file.
