+++ pyproject.toml
@@ -1,0 +1,19 @@
+[tool.poetry]
+name = "text-repetetive"
+version = "0.1.0"
+description = ""
+authors = ["laundmo <laurinschmidt2001@gmail.com>"]
+
+[tool.poetry.dependencies]
+python = "^3.10"
+Pillow = "^9.2.0"
+
+[tool.poetry.dev-dependencies]
+pytest = "^5.2"
+
+[tool.poetry.scripts]
+generate = "text_repetetive.__main__:main"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
